fix(web): refresh cached track lists after a scene merge
A merge rewrites the generated `genre` scalar on every affected row, but the
panel invalidated only the genres endpoint, so Browse and Tracks kept showing
the retired spelling until they remounted. Invalidate ['library','rows'] — the
family every cached list of Tracks sits under, the same reach a tag edit or a
block re-stamp uses — plus coverage, whose byGenre tally the merge just
invalidated server-side.

The button's source list also filtered case-insensitively, which disabled the
merge on exactly the case-duplicate tail the section exists to clean up now
that the server treats those as a real merge. It matches the server verbatim.

Also: SceneCount/SceneAlias move to library/types.ts beside the other shared
row types; the merge toast distinguishes "rule recorded for the next scan"
from "nothing to do at all"; and the folds list says its left-hand side is the
fold key rather than leaving a lower-cased tag looking like a listing entry
that matches nothing above it.
